move random port to core module
