+++ apm-agent-core/src/test/java/co/elastic/apm/agent/testutils/TestPort.java
@@ -11,9 +11,9 @@
  * the Apache License, Version 2.0 (the "License"); you may
  * not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -22,7 +22,7 @@
  * under the License.
  * #L%
  */
-package co.elastic.apm.agent.grpc;
+package co.elastic.apm.agent.testutils;
 
 import javax.net.ServerSocketFactory;
 import java.net.InetAddress;
@@ -36,6 +36,9 @@ public class TestPort {
     private static final int MIN_PORT = 1024;
     private static final int MAX_PORT = 65536;
 
+    /**
+     * @return random available TCP port
+     */
     public static int getAvailableRandomPort() {
         int port;
         do {
