+++ apm-agent-plugins/apm-grpc/apm-grpc-plugin/src/test/java/co/elastic/apm/agent/grpc/GrpcTest.java
@@ -26,6 +26,7 @@ package co.elastic.apm.agent.grpc;
 
 import co.elastic.apm.agent.grpc.testapp.GrpcApp;
 import co.elastic.apm.agent.grpc.testapp.GrpcAppProvider;
+import co.elastic.apm.agent.testutils.TestPort;
 
 public class GrpcTest {
